chore fix and sqlc gen files

+++ migrations/20260423113336_init.sql
@@ -46,8 +46,8 @@ CREATE TABLE product (
     id          SERIAL      PRIMARY KEY,
     name        VARCHAR     NOT NULL,
     price       INT,
-    weight      FLOAT,
-    volume      FLOAT,
+    weight      INT,
+    volume      INT,
     created_at  TIMESTAMP   NOT NULL DEFAULT NOW()
 );
